+++ pkg/security/probe/hookpoints.go
@@ -8,20 +8,14 @@
 package probe
 
 import (
+	"github.com/DataDog/datadog-agent/pkg/security/ebpf"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/eval"
 )
-
-// KProbe describes a Linux Kprobe
-type KProbe struct {
-	Name      string
-	EntryFunc string
-	ExitFunc  string
-}
 
 // HookPoint represents
 type HookPoint struct {
 	Name       string
-	KProbes    []*KProbe
+	KProbes    []*ebpf.KProbe
 	Tracepoint string
 	Optional   bool
 	EventTypes []eval.EventType
@@ -30,7 +24,7 @@ type HookPoint struct {
 var allHookPoints = []*HookPoint{
 	{
 		Name: "security_inode_setattr",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/security_inode_setattr",
 		}},
 		EventTypes: []eval.EventType{"chmod", "chown", "utimes"},
@@ -72,14 +66,14 @@ var allHookPoints = []*HookPoint{
 	},
 	{
 		Name: "mnt_want_write",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/mnt_want_write",
 		}},
 		EventTypes: []eval.EventType{"utimes", "chmod", "chown", "rmdir", "unlink", "rename"},
 	},
 	{
 		Name: "mnt_want_write_file",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/mnt_want_write_file",
 		}},
 		EventTypes: []eval.EventType{"chown"},
@@ -106,14 +100,14 @@ var allHookPoints = []*HookPoint{
 	},
 	{
 		Name: "vfs_mkdir",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/vfs_mkdir",
 		}},
 		EventTypes: []eval.EventType{"mkdir"},
 	},
 	{
 		Name: "filename_create",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/filename_create",
 		}},
 		EventTypes: []eval.EventType{"mkdir", "link"},
@@ -130,7 +124,7 @@ var allHookPoints = []*HookPoint{
 	},
 	{
 		Name: "vfs_rmdir",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/vfs_rmdir",
 		}},
 		EventTypes: []eval.EventType{"rmdir", "unlink"},
@@ -142,7 +136,7 @@ var allHookPoints = []*HookPoint{
 	},
 	{
 		Name: "vfs_rename",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/vfs_rename",
 		}},
 		EventTypes: []eval.EventType{"rename"},
@@ -164,7 +158,7 @@ var allHookPoints = []*HookPoint{
 	},
 	{
 		Name: "vfs_link",
-		KProbes: []*KProbe{{
+		KProbes: []*ebpf.KProbe{{
 			EntryFunc: "kprobe/vfs_link",
 		}},
 		EventTypes: []eval.EventType{"link"},
